comit befor pull

+++ src/main/java/com/stackhack/taskmanagement/repo/CustomerRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.stackhack.taskmanagement.entity.Customer;
 
-public interface CustomerRepo extends JpaRepository<Customer, Integer> {
+public interface CustomerRepo extends JpaRepository<Customer, Long> {
 
 }
